services/discord_bot/requirements.txt
discord.py==2.3.2
redis==5.0.1
aiohttp==3.9.1